+++ agent/api/container.go
@@ -136,7 +136,14 @@ type Container struct {
 	// metadata file
 	MetadataFileUpdated bool `json:"metadataFileUpdated"`
 
-	knownExitCode     *int
+	// KnownExitCodeUnsafe specifies the exit code for the container.
+	// It is exposed outside of the package so that it's marshalled/unmarshalled in
+	// the JSON body while saving the state.
+	// NOTE: Do not access KnownExitCodeUnsafe directly. Instead, use `GetKnownExitCode`
+	// and `SetKnownExitCode`.
+	KnownExitCodeUnsafe *int `json:"KnownExitCode"`
+
+	// KnownPortBindings is an array of port bindings for the container.
 	KnownPortBindings []PortBinding
 
 	// SteadyStateStatusUnsafe specifies the steady state status for the container
@@ -244,14 +251,16 @@ func (c *Container) SetSentStatus(status ContainerStatus) {
 func (c *Container) SetKnownExitCode(i *int) {
 	c.lock.Lock()
 	defer c.lock.Unlock()
-	c.knownExitCode = i
+
+	c.KnownExitCodeUnsafe = i
 }
 
 // GetKnownExitCode returns the container exit code
 func (c *Container) GetKnownExitCode() *int {
 	c.lock.RLock()
 	defer c.lock.RUnlock()
-	return c.knownExitCode
+
+	return c.KnownExitCodeUnsafe
 }
 
 // SetRegistryAuthCredentials sets the credentials for pulling image from ECR
@@ -370,7 +379,7 @@ func (c *Container) IsEssential() bool {
 	return c.Essential
 }
 
-// LogAuthExecutionRole returns true if the auth is by exectution role
+// AWSLogAuthExecutionRole returns true if the auth is by execution role
 func (c *Container) AWSLogAuthExecutionRole() bool {
 	return c.LogsAuthStrategy == awslogsAuthExecutionRole
 }